Split the BT leaves one per file and share what they had in common
