+++ workspace/src/g1_orchestration/CMakeLists.txt
@@ -15,9 +15,21 @@ find_package(lifecycle_msgs REQUIRED)
 find_package(tf2 REQUIRED)
 find_package(tf2_geometry_msgs REQUIRED)
 
+# One file per leaf, under src/skills, the layout nav2_behavior_tree uses. Listed rather than
+# globbed so a file that fails to get added fails the build instead of vanishing silently.
 add_library(g1_bt_nodes
-  src/skill_nodes.cpp
   src/arm_authority.cpp
+  src/port_types.cpp
+  src/registration.cpp
+  src/service_leaf.cpp
+  src/skills/approach_object.cpp
+  src/skills/arm_authority_leaves.cpp
+  src/skills/clear_costmaps.cpp
+  src/skills/navigate_to_pose.cpp
+  src/skills/pick.cpp
+  src/skills/place.cpp
+  src/skills/retreat.cpp
+  src/skills/set_arm_posture.cpp
 )
 target_include_directories(g1_bt_nodes PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
@@ -31,9 +43,14 @@ ament_target_dependencies(g1_bt_nodes PUBLIC
 add_executable(g1_bt_executor src/g1_bt_executor_main.cpp)
 target_link_libraries(g1_bt_executor PUBLIC g1_bt_nodes)
 
+# Regenerates the Groot2 palette. Not run by the build: the model is checked in so Groot2 can
+# open the project straight from a clone, and test_node_model fails if the two drift.
+add_executable(g1_bt_node_model src/g1_bt_node_model_main.cpp)
+target_link_libraries(g1_bt_node_model PUBLIC g1_bt_nodes)
+
 install(DIRECTORY include/ DESTINATION include)
 install(TARGETS g1_bt_nodes EXPORT export_${PROJECT_NAME} DESTINATION lib)
-install(TARGETS g1_bt_executor DESTINATION lib/${PROJECT_NAME})
+install(TARGETS g1_bt_executor g1_bt_node_model DESTINATION lib/${PROJECT_NAME})
 install(DIRECTORY launch trees DESTINATION share/${PROJECT_NAME})
 
 ament_export_targets(export_${PROJECT_NAME} HAS_LIBRARY_TARGET)
@@ -50,6 +67,12 @@ if(BUILD_TESTING)
   ament_add_gmock(test_tree_loads test/test_tree_loads.cpp)
   target_link_libraries(test_tree_loads g1_bt_nodes)
   target_compile_definitions(test_tree_loads PRIVATE
+    G1_TREES_DIR="${CMAKE_CURRENT_SOURCE_DIR}/trees")
+
+  # A stale Groot2 palette invites trees that cannot load.
+  ament_add_gmock(test_node_model test/test_node_model.cpp)
+  target_link_libraries(test_node_model g1_bt_nodes)
+  target_compile_definitions(test_node_model PRIVATE
     G1_TREES_DIR="${CMAKE_CURRENT_SOURCE_DIR}/trees")
 
   # The acquire sequence against g1_bringup's script, which is the other implementation of it.
