+++ workspace/src/g1_orchestration/include/g1_orchestration/service_leaf.hpp
@@ -1,0 +1,67 @@
+#ifndef G1_ORCHESTRATION__SERVICE_LEAF_HPP_
+#define G1_ORCHESTRATION__SERVICE_LEAF_HPP_
+
+/**
+ * @file service_leaf.hpp
+ * @brief Blocking service calls from a leaf, on a node the executor does not own.
+ *
+ * spin_until_future_complete on a node an executor already holds throws rather than waiting, so
+ * these calls need a node of their own. It is created for the call and destroyed with it.
+ *
+ * Two leaves grew their own copy of this before it was hoisted here.
+ */
+
+#include <behaviortree_cpp/action_node.h>
+
+#include <chrono>
+#include <memory>
+#include <rclcpp/rclcpp.hpp>
+#include <string>
+#include <utility>
+
+#include "g1_orchestration/ros_action_node.hpp"
+
+namespace g1_orchestration
+{
+
+/// A node nothing else is spinning, for the duration of one call sequence.
+rclcpp::Node::SharedPtr makeClientNode(const std::string& name);
+
+template <typename ServiceT>
+typename ServiceT::Response::SharedPtr callService(
+    const rclcpp::Node::SharedPtr& node, const std::string& service,
+    const typename ServiceT::Request::SharedPtr& request, double timeout_s)
+{
+    auto client = node->create_client<ServiceT>(service);
+    if (!client->wait_for_service(std::chrono::duration<double>(timeout_s)))
+    {
+        RCLCPP_ERROR(node->get_logger(), "service '%s' never appeared", service.c_str());
+        return nullptr;
+    }
+    auto future = client->async_send_request(request);
+    if (rclcpp::spin_until_future_complete(node, future, std::chrono::duration<double>(timeout_s)) !=
+        rclcpp::FutureReturnCode::SUCCESS)
+    {
+        RCLCPP_ERROR(node->get_logger(), "call to '%s' did not return", service.c_str());
+        return nullptr;
+    }
+    return future.get();
+}
+
+/// Base for leaves that complete within one tick. Holds the tree's node for logging; the calls
+/// themselves use makeClientNode().
+class ServiceLeaf : public BT::SyncActionNode
+{
+public:
+    ServiceLeaf(const std::string& name, const BT::NodeConfig& config, RosContext context)
+      : BT::SyncActionNode(name, config)
+      , node_(std::move(context.node))
+    {}
+
+protected:
+    rclcpp::Node::SharedPtr node_;
+};
+
+}  // namespace g1_orchestration
+
+#endif  // G1_ORCHESTRATION__SERVICE_LEAF_HPP_
