+++ workspace/src/g1_orchestration/include/g1_orchestration/skills/set_arm_posture.hpp
@@ -1,0 +1,25 @@
+#ifndef G1_ORCHESTRATION__SKILLS__SET_ARM_POSTURE_HPP_
+#define G1_ORCHESTRATION__SKILLS__SET_ARM_POSTURE_HPP_
+
+#include <g1_msgs/action/set_arm_posture.hpp>
+#include <string>
+
+#include "g1_orchestration/skill_action_node.hpp"
+
+namespace g1_orchestration
+{
+
+/// Moves a planning group to one of its named SRDF poses.
+class SetArmPosture : public SkillActionNode<g1_msgs::action::SetArmPosture>
+{
+public:
+    SetArmPosture(const std::string& name, const BT::NodeConfig& config, RosContext context);
+    static BT::PortsList providedPorts();
+
+protected:
+    bool fillGoal(Goal& goal) override;
+};
+
+}  // namespace g1_orchestration
+
+#endif  // G1_ORCHESTRATION__SKILLS__SET_ARM_POSTURE_HPP_
